delete webhook when project is deleted
