+++ pkg/microservice/aslan/core/common/service/cronjob.go
@@ -17,16 +17,7 @@ limitations under the License.
 package service
 
 import (
-	"encoding/json"
-
-	"go.uber.org/zap"
-
-	"github.com/koderover/zadig/pkg/microservice/aslan/config"
 	"github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/models"
-	"github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/mongodb"
-	"github.com/koderover/zadig/pkg/microservice/aslan/core/common/service/nsq"
-	"github.com/koderover/zadig/pkg/setting"
-	e "github.com/koderover/zadig/pkg/tool/errors"
 )
 
 type CronjobPayload struct {
@@ -36,26 +27,4 @@ type CronjobPayload struct {
 	JobType     string             `json:"job_type"`
 	DeleteList  []string           `json:"delete_list,omitempty"`
 	JobList     []*models.Schedule `json:"job_list,omitempty"`
-}
-
-func RemoveCronjob(workflowName string, log *zap.SugaredLogger) error {
-	err := mongodb.NewCronjobColl().Delete(&mongodb.CronjobDeleteOption{ParentName: workflowName, ParentType: config.WorkflowCronjob})
-	if err != nil {
-		// FIXME: HOW TO DO THIS
-		log.Errorf("Failed to delete %s 's cronjob, the error is: %v", workflowName, err)
-		//return e.ErrDeleteWorkflow.AddDesc(err.Error())
-	}
-	payload := CronjobPayload{
-		Name:    workflowName,
-		JobType: config.WorkflowCronjob,
-		Action:  setting.TypeDisableCronjob,
-	}
-	pl, _ := json.Marshal(payload)
-	err = nsq.Publish(config.TopicCronjob, pl)
-	if err != nil {
-		log.Errorf("Failed to publish to nsq topic: %s, the error is: %v", config.TopicCronjob, err)
-		return e.ErrUpsertCronjob.AddDesc(err.Error())
-	}
-
-	return nil
 }
